utils.c: fix implicit declaration of strcasestr()
The strcasestr() function is a GNU extension. Added a header to request
GNU extensions for main.c. This will help indicate developers that
portability is constrained to GNU systems that use GNU libc.

The proper fix, to let main.c compile on non-GNU systems is to drop
the use of strcasestr().

Reference:
- http://linux.die.net/man/3/strcasestr